requirements.txt
scikit-video 
scikit-learn
tensorflow
imutils 
opencv-python
SoccerNet 
moviepy 
nvidia-tensorrt
ffmpy
m3u8
streamlink
pyyaml
git+https://github.com/openai/whisper.git
